sdk: propagate body-merge errors, widen SizeBytes to int64
Fixes the two remaining low-severity items from the review.

- clients/go/request: mergeBody silently swallowed json.Marshal/Unmarshal
  errors in the struct-body default case, so a body that failed to encode (or
  wasn't a JSON object) was dropped and the request went out missing the
  caller's body. Return the error and propagate it from newRequest.
- clients/go/files: FilesCreateUploadParams.SizeBytes was `int`, which is
  32-bit on 32-bit builds — a >2 GiB upload size would overflow (and the CLI
  cast int(size) truncated an int64 flag even before that). Widen to int64 and
  drop the lossy casts at all four call sites (cli files create/upload, plus
  the build-tagged oagen overlays).

Adds a mergeBody regression test. Both modules build/vet/test clean; the
generated oagen build config (retab_oagen_cli_prototype + _files) compiles.

+++ cli/cmd/files.go
@@ -263,7 +263,7 @@ func uploadFile(ctx context.Context, client *retab.Client, uploadPath string) (*
 	prepared, err := client.Files.CreateUpload(ctx, &retab.FilesCreateUploadParams{
 		Filename:    filename,
 		ContentType: &contentType,
-		SizeBytes:   len(data),
+		SizeBytes:   int64(len(data)),
 		Sha256:      &sha256Hash,
 	})
 	if err != nil {
@@ -726,7 +726,7 @@ Steps: (1) ` + "`create-upload`" + ` returns ` + "`{id, upload_url, ...}`" + `;
 		result, err := client.Files.CreateUpload(ctx, &retab.FilesCreateUploadParams{
 			Filename:    filename,
 			ContentType: &contentType,
-			SizeBytes:   int(size),
+			SizeBytes:   size,
 			Sha256:      &sha256Hash,
 		})
 		if err != nil {
